refactor(bpf): move ICMPv6 packet generation to a separate file
The goal is to make these functions reusable for the ICMPv6 policy
denial feature, so moving them to a shared icmp6.h file
